# configs/swaggerConfig.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cloud Scrapy API",
    "version": "0.0.1",
    "description": "API to interact with headless browser engine bot on the cloud to scrape data",
    "license": {
      "name": "MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Lucual18",
      "url": "https://www.lucual18.com",
      "email": "layusoelu@lucual18.es"
    }
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-Key",
      "description": "Cloud Scrapy key to authenticate"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/files/public/screenshots/{imageName}": {
      "get": {
        "tags": [
          "Files"
        ],
        "description": "Download an image passing the exact name",
        "parameters": [
          {
            "name": "imageName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/files/public/logs/{logName}": {
      "get": {
        "tags": [
          "Files"
        ],
        "description": "Download a log file passing the exact name",
        "parameters": [
          {
            "name": "logName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/api/files/public/pdfs/{pdfName}": {
      "get": {
        "tags": [
          "Files"
        ],
        "description": "Download a pdf file passing the exact name",
        "parameters": [
          {
            "name": "pdfName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
}